[chore] moving interfaces and linting
