+++ src/main/index.ts
@@ -1,23 +1,3 @@
-import {
-  IPuppeteerPoolConfig,
-  createPuppeteerPool,
-  USE_COUNT,
-  ID,
-} from './puppeteer-pool';
-import {
-  IDimensions,
-  IConfig,
-  Svg2png,
-  svg2png,
-} from './svg2png';
-
-export {
-  IPuppeteerPoolConfig,
-  createPuppeteerPool,
-  USE_COUNT,
-  ID,
-  IDimensions,
-  IConfig,
-  Svg2png,
-  svg2png,
-};
+export * from './interfaces';
+export * from './puppeteer-pool';
+export * from './svg2png';
