+++ src/main/svg2png-cli.ts
@@ -2,6 +2,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import * as yargs from 'yargs';
+
 import { svg2png } from './svg2png';
 
 async function main() {
