+++ src/main/svg2png.ts
@@ -1,54 +1,10 @@
-import { Page, Browser } from 'puppeteer';
-import { Pool } from 'generic-pool';
-import * as sharp from 'sharp';
-import { createPuppeteerPool } from './puppeteer-pool';
 import fileUrl = require('file-url');
+import { Pool } from 'generic-pool';
+import { Browser, Page } from 'puppeteer';
+import * as sharp from 'sharp';
 
-/**
- * An interface to help us specify the dimensions of an svg object.
- */
-interface IDimensions {
-  /**
-   * The width of the svg. In pixels.
-   */
-  width: number;
-
-  /**
-   * The height of the svg. In pixels.
-   */
-  height: number;
-
-  /**
-   * A scale factor to expand or shrink the svg.
-   */
-  scale: number;
-}
-
-/**
- * A configuration interface to use in the `svg2png` function.
- */
-interface IConfig extends Partial<IDimensions> {
-  /**
-   * The url of the svg to convert.
-   */
-  url: string;
-
-  /**
-   * A limit on the allowed time to load the svg in the browser.
-   */
-  navigationTimeout?: number;
-
-  /**
-   * A limit of the allowed time to spend on the conversion.
-   */
-  conversionTimeout?: number;
-
-  /**
-   * If `true`, console messages will be printed specifying the actions taken during the
-   * conversion.
-   */
-  debug?: boolean;
-}
+import { IConfig, IDimensions } from './interfaces';
+import { createPuppeteerPool } from './puppeteer-pool';
 
 /**
  * The possible status of a conversion.
