+++ src/test/index.ts
@@ -1,7 +1,6 @@
 import { Svg2png } from '../main/svg2png';
-
-import './puppeteer-pool';
 import './conversion-tests';
+import './puppeteer-pool';
 
 describe('Closing', () => {
   // Using the library creates a pool for Svg2png. Thus we need to make sure to close it at
